Refactored Selenium Tests
