+++ requirements.txt
@@ -15,7 +15,7 @@ Jinja2==3.1.6
 Mako==1.3.10
 MarkupSafe==3.0.2
 narwhals==1.38.0
-numpy==2.2.5
+numpy==2.0.2
 packaging==25.0
 pandas==2.2.3
 plotly==6.0.1
